requirements.txt
openai
streamlit
gtts
PyMuPDF
python-docx
pandas
